[move] Refactor `GasSchedule` to `move_core_types`
