+++ language/move-vm/types/src/native_functions/hash.rs
@@ -8,12 +8,10 @@ use crate::{
 };
 use libra_crypto::HashValue;
 use libra_types::vm_error::{StatusCode, VMStatus};
+use move_core_types::gas_schedule::{CostTable, NativeCostIndex};
 use sha2::{Digest, Sha256};
 use std::collections::VecDeque;
-use vm::{
-    errors::VMResult,
-    gas_schedule::{CostTable, NativeCostIndex},
-};
+use vm::errors::VMResult;
 
 pub fn native_sha2_256(
     _ty_args: Vec<Type>,
