+++ language/vm/src/transaction_metadata.rs
@@ -1,11 +1,13 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::gas_schedule::{AbstractMemorySize, GasAlgebra, GasCarrier, GasPrice, GasUnits};
 use libra_crypto::{ed25519::Ed25519PrivateKey, PrivateKey};
 use libra_types::{
     account_address::AccountAddress,
     transaction::{authenticator::AuthenticationKeyPreimage, SignedTransaction},
+};
+use move_core_types::gas_schedule::{
+    AbstractMemorySize, GasAlgebra, GasCarrier, GasPrice, GasUnits,
 };
 use std::{convert::TryFrom, time::Duration};
